+++ modules/stores/social/circles/controller/circles.controller.ts
@@ -1,7 +1,24 @@
+import validate from "../../../validate.ts";
 import circlesService from "../service/circles.service.ts";
 
 class CirclesController {
     async queryCircles(ctx: any): Promise<any> {
+        const body = await ctx.request.body().value;
+
+        const result = await validate.validateUser(body["user_id"])
+
+        if (!result) {
+            ctx.response.status = 401;
+            ctx.response.body = {
+                meta: {
+                    code: 401,
+                    status: "Not Authorized",
+                }
+            }
+
+            return;
+        }
+        
         ctx.response.status = 200;
         ctx.response.body = {
             meta: {
@@ -13,6 +30,22 @@ class CirclesController {
     }
 
     async getCirclesOfUser(ctx: any): Promise<any> {
+        const body = await ctx.request.body().value;
+
+        const result = await validate.validateUser(body["user_id"])
+
+        if (!result) {
+            ctx.response.status = 401;
+            ctx.response.body = {
+                meta: {
+                    code: 401,
+                    status: "Not Authorized",
+                }
+            }
+
+            return;
+        }
+
         ctx.response.status = 200;
         ctx.response.body = {
             meta: {
@@ -24,6 +57,22 @@ class CirclesController {
     }
     
     async getCircleById(ctx: any): Promise<any> {
+        const body = await ctx.request.body().value;
+
+        const result = await validate.validateUser(body["user_id"])
+
+        if (!result) {
+            ctx.response.status = 401;
+            ctx.response.body = {
+                meta: {
+                    code: 401,
+                    status: "Not Authorized",
+                }
+            }
+
+            return;
+        }
+
         ctx.response.status = 200;
         ctx.response.body = {
             meta: {
@@ -37,6 +86,20 @@ class CirclesController {
     async getUsersOfCircle(ctx: any): Promise<any> {
         const body = await ctx.request.body().value;
 
+        const result = await validate.validateUser(body["user_id"])
+
+        if (!result) {
+            ctx.response.status = 401;
+            ctx.response.body = {
+                meta: {
+                    code: 401,
+                    status: "Not Authorized",
+                }
+            }
+
+            return;
+        }
+
         ctx.response.status = 200;
         ctx.response.body = {
             meta: {
@@ -48,6 +111,22 @@ class CirclesController {
     }
 
     async createCircleConnection(ctx: any): Promise<any> {
+        const body = await ctx.request.body().value;
+
+        const result = await validate.validateUser(body["user_id"])
+
+        if (!result) {
+            ctx.response.status = 401;
+            ctx.response.body = {
+                meta: {
+                    code: 401,
+                    status: "Not Authorized",
+                }
+            }
+
+            return;
+        }
+
         await circlesService.createCircleConnection(ctx)
 
         ctx.response.status = 201;
@@ -60,6 +139,22 @@ class CirclesController {
     }
 
     async createCircle(ctx: any): Promise<any> {
+        const body = await ctx.request.body().value;
+
+        const result = await validate.validateUser(body["user_id"])
+
+        if (!result) {
+            ctx.response.status = 401;
+            ctx.response.body = {
+                meta: {
+                    code: 401,
+                    status: "Not Authorized",
+                }
+            }
+
+            return;
+        }
+
         await circlesService.createCircle(ctx)
 
         ctx.response.status = 201;
@@ -72,6 +167,22 @@ class CirclesController {
     }
 
     async deleteCircle(ctx: any): Promise<any> {
+        const body = await ctx.request.body().value;
+
+        const result = await validate.validateUser(body["user_id"])
+
+        if (!result) {
+            ctx.response.status = 401;
+            ctx.response.body = {
+                meta: {
+                    code: 401,
+                    status: "Not Authorized",
+                }
+            }
+
+            return;
+        }
+        
         await circlesService.deleteCircle(ctx);
 
         ctx.response.status = 200;
